Change (1)
While starting off with the code I started to think use the code "Tkinter". The reason why I use this code is because tkinter is a lightweight library included with Python, ideal for simple GUI applications. Now, I thought that i could use the code "Random" due to generating an unpredictable number for the guessing game.