Change 2
After figuring out what should be next on my code,  I use the code "attempts" and "answers" to  keep track of the number of guesses and generate a random target number. Now, why use these? because 1.) add challenges by adding a limit attempt and 2.) ensure fairness and unprecedented in the game