Change 3
While I was stuck on what to do of this code (change 1-2) that I was a bit to quit but continue to make some changes. So I started to use some coding variables where to make the game function. Such as, the "check_answer" function makes the players guess at the remaining attempt. Compare input to answer by displaying a winning message towards the player by hiding the "check" button. Finally, at line 26-27 I started to make sure that non-numbers input that does not make the program crash out as my previous (beta version) and let the user to try again. Now, the reason why I added these changes because for starters I do not want this main code to crash out as my previous code (beta version). Additionally, the coding structure summarize the errors correctly, making sure that the user playing the game has a non-crash experience.