Change 4
While thinking for 2 hours I manage to find the right functions that could possibly work with the number guessing game. I started to use "root=tk()" to start with the features and details about the game and how is going to look like. However, this part is still in the works it's not finish and I need to add some more coding function to fix line 17 through 25.