Change 5
After an hour, thinking what to be putting on the game, I starting to find some inspiration of what a game needs. Now, I found that the mid-beginning lines (line 16-25), there were some missing coding that I need to fix in order to stop the error. So, I proceeded to add some coding function on line (37-46). A quit and check button weather the player want to quit or check their guess. Finally, I added some hints because a game need hints on a guessing game. Now, the reason why I added these code functions in my guessing game is to have a layout where buttons clearly function, a text field for input variables, and labels for adding details to ensure an interface.